TLE/RE: Add Time and Memory Limit to each request
Timeout set to 3s

Fixes: #3

+++ Util/Del.py
@@ -21,7 +21,8 @@ from time import time
 def delete(source_file, binary_file, input_file, output_file, err_file):
     try:
         os.remove(source_file)
-        os.remove(binary_file)
+        if binary_file:
+            os.remove(binary_file)
         if input_file:
             os.remove(input_file)
         os.remove(output_file)
